Fix document table filter

+++ gwt/document/src/main/java/ru/protei/portal/ui/document/client/activity/table/AbstractDocumentTableView.java
@@ -9,7 +9,7 @@ public interface AbstractDocumentTableView extends IsWidget {
 
     void clearRecords();
 
-    void setRecordCount(Long count);
+    void setRecordCount(int count);
 
     int getPageSize();
